# DAOS-11409 mgmt: Improve pool create, startup time with md_on_ssd