--- a/src/mgmt/srv_internal.h
+++ b/src/mgmt/srv_internal.h
@@ -37,6 +37,7 @@ void ds_mgmt_profile_hdlr(crt_rpc_t *rpc);
 void ds_mgmt_pool_get_svcranks_hdlr(crt_rpc_t *rpc);
 void ds_mgmt_pool_find_hdlr(crt_rpc_t *rpc);
 void ds_mgmt_mark_hdlr(crt_rpc_t *rpc);
+void dss_bind_to_xstream_cpuset(int tgt_id);
 
 /** srv_system.c */
 /* Management service (used only for map broadcast) */