@sherintg 2023-02-22
> The commit comment doesn't mention how much time we can save by parallelizing tmpfs fallocate, I'm just wondering if you've measured the improvement from this specific change?

Please find the performance improvement details below:
Pool create command run: **dmg pool create -s 50G -n 500G -r 1 pool1**
Number of targets per rank: **8**

**Without the fix** dmg pool create took **37.6 seconds** to complete. Of this **8.2 seconds** was attributed to fallocate() and **3.6 * 8 seconds** was attributed to pool open.

**With the fix** pool create takes **5.6 seconds** to complete. Of this **1.1 seconds** is attributed to parallel fallocate (8 threads) and  each parallel pool open took **4.1 seconds** to complete.