@sherintg 2023-02-22
> I wonder if the removal of the lock should be done on master branch instead?

Please let me the know, if I have to create a PR to the master branch with change alone.